refactor: reduce code duplication on pages
