+++ src/pages/pupils/beta/programmes/[programmeSlug]/options-l.tsx
@@ -2,18 +2,15 @@
 // http://localhost:3000/pupils/beta/programmes/maths-secondary-year-10/options-l
 // CF http://localhost:3000/pupils/beta/programmes/maths-secondary-year-10/options
 
-import { GetStaticPathsResult, GetStaticProps } from "next";
+import { GetStaticProps } from "next";
 
-import {
-  shouldSkipInitialBuild,
-  getFallbackBlockingConfig,
-} from "@/node-lib/isr";
 import { PupilViewsProgrammeListing } from "@/components/PupilViews/PupilProgrammeListing/PupilProgrammeListing.view";
 import {
   ProgrammesPageProps,
   OptionsURLParams,
   getPupilOptionData,
 } from "@/pages-helpers/pupil/options-pages/options-pages-helpers";
+import { getStaticPaths as getStaticPathsTemplate } from "@/pages-helpers/get-static-paths";
 
 const ProgrammesPage = ({
   programmes,
@@ -30,17 +27,7 @@ const ProgrammesPage = ({
   );
 };
 
-export const getStaticPaths = async () => {
-  if (shouldSkipInitialBuild) {
-    return getFallbackBlockingConfig();
-  }
-
-  const config: GetStaticPathsResult<OptionsURLParams> = {
-    fallback: "blocking",
-    paths: [],
-  };
-  return config;
-};
+export const getStaticPaths = getStaticPathsTemplate<OptionsURLParams>;
 
 export const getStaticProps: GetStaticProps<
   ProgrammesPageProps,
