+++ src/pages/pupils/beta/programmes/[programmeSlug]/options.tsx
@@ -3,13 +3,9 @@
 // http://localhost:3000/pupils/beta/programmes/maths-secondary-year-10/options
 // http://localhost:3000/pupils/beta/programmes/maths-secondary-year-11/options (should 404)
 
-import { GetStaticPathsResult, GetStaticProps } from "next";
+import { GetStaticProps } from "next";
 import { useSearchParams } from "next/navigation";
 
-import {
-  shouldSkipInitialBuild,
-  getFallbackBlockingConfig,
-} from "@/node-lib/isr";
 import { PupilViewsProgrammeListing } from "@/components/PupilViews/PupilProgrammeListing/PupilProgrammeListing.view";
 import {
   ProgrammesPageProps,
@@ -18,6 +14,7 @@ import {
   getPupilOptionData,
 } from "@/pages-helpers/pupil/options-pages/options-pages-helpers";
 import OakError from "@/errors/OakError";
+import { getStaticPaths as getStaticPathsTemplate } from "@/pages-helpers/get-static-paths";
 
 const ProgrammesPage = ({
   programmes,
@@ -46,17 +43,7 @@ const ProgrammesPage = ({
   );
 };
 
-export const getStaticPaths = async () => {
-  if (shouldSkipInitialBuild) {
-    return getFallbackBlockingConfig();
-  }
-
-  const config: GetStaticPathsResult<OptionsURLParams> = {
-    fallback: "blocking",
-    paths: [],
-  };
-  return config;
-};
+export const getStaticPaths = getStaticPathsTemplate<OptionsURLParams>;
 
 export const getStaticProps: GetStaticProps<
   ProgrammesPageProps,
